[dg docs] Iteration on JSONSchema (#28446)
## Summary & Motivation

Some tweaks to the nested JSONSchema display, including:

- Move the schema back above the example
- Better determination of expandability
- Fix overflow and wrapping for smaller viewports

Still a little concerned about pathological nesting cases here and
various flavors of schema that this might not handle fully yet, but it
seems decent for the current test case.


https://github.com/user-attachments/assets/581802ec-c685-4c08-a69b-3d56fb01eaed


Would like to try this on more examples.

## How I Tested These Changes

dg docs, do some expanding/collapsing on the schema view.