[dagster-graphql] Allow for basic auth (#28038)
Allow propagating HTTPBasicAuth instance to RequestsHTTPTransport for
basic authentication

## Summary & Motivation
For allowing requests to Dagster instances deployed on docker or K8s
with nginx-htpasswd in front.
 
## How I Tested These Changes
I tested these changes by patching the library in my virtual
environment, and invoking the Dagster GraphQL API to reload a code
location in our UAT Dagster instance.
 
## Changelog

[dagster-graphql] Extend the DagsterGraphQLClient constructor with the `auth` keyword
argument, that is in turn passed along to the RequestsHTTPTransport
constructor