[ui] Handle intermediate states after creating dynamic partitions (#28514)
## Summary & Motivation

After you create a dynamic partition, asset health data refreshes but
some assets data may be pulled from cache while other data is fresh.
This creates a scenario where your selection of assets have inconsistent
number of partitions.

We've been supporting this by sidestepping an assertion in the code.
This PR goes a step further and adds code + tests to clarify the desired
behavior in this case and removes the assertion.

To test this I simulated the asset health data loading very slowly and
noticed a separate bug where smashing Launch Run after creating a new
dynamic partition would send `null` as the partitionName if none of the
assets' health info had refreshed. I made the code detect that this is
happening and briefly disable the Launch button. We could alternatively
display a "please wait a moment" modal if the user clicks during this
period, but it'd be unclear how long to wait before clicking it again.

## How I Tested These Changes

Tested with a dynamic job containing multiple assets + simulated slow
asset health responses.

## Changelog

[ui] The Asset > Partitions page no longer displays an error in some
cases when creating dynamic partitions.
[ui] The Launch and Report Events buttons no longer error if you click
it immediately after creating a new dynamic partition.

Co-authored-by: bengotow <bgotow@elementl.com>