[da] Create `executed_with_tags` condition (#27684)
## Summary & Motivation

This condition revives the functionality from the long-dead `AutoMaterializeAssetPartitionsFilter` of the `AutoMaterializeRule`.

This rule was valuable in that it made it possible to filter out events from certain runs from being considered. However, it was painfully over-specified and had bugs related to the fact that it was re-evaluated every tick, always re-fetching the latest materialization event.

The new DA system does not have such limitations, and this PR does not introduce any new query paths which are not already exercised by the existing `AutomationCondition.eager()`, which already fetches the run associated with the most recent update to determine if that run contains the asset we're targeting or not (this is the `executed_with_root_target` condition).

Given that we already have this run information in memory, I think it makes a lot of sense to be able to interrogate this run information more deeply, particularly investigating the tags associated with it.

The main usecase here is to filter out runs that were _not_ executed via the automation system from impacting downstream compute, e.g. if someone does a manual run of an asset, sometimes you don't want that to impact your eager condition.

## How I Tested These Changes

## Changelog

Added a new `AutomationCondition.executed_with_tags()` condition that makes it possible to only listen to updates that happened with runs with particular tags.