Explicitly pass through BUILDKITE_ANALYTICS_TOKEN
I think what's happening here is that the pre-command hook is running,
but the docker command was already pre-generated in the initial generate
pipeline step.

I don't love this pattern of passing everything through - as we converge
our OSS and internal build tooling, presumably that'll end up on the
chopping block.

But for now, I'm just appending the env I expect to exist so it gets set
on the docker container.