Inject env vars in dagster-components list definitions
Summary:
To be fully compliant with the env vars API, requires the CLi command knowing what location it is. I'm a little unclear how to add new options like this in a back-compat way though, given that we can't be certain that dg won't be on a newer version than the dagster-components package version that it calls into. Has that class of issues come up yet?

> Insert changelog entry or delete this section.

test

> Insert changelog entry or delete this section.