[dg] Update newly_requested to allow invocation in deps context (#28569)
## Summary & Motivation

Previously, we used the cursor of the root context to determine if a given entity was newly requested, which doesn't work in cases where the asset key is not the same as they of the root target.

This stashes the full daemon cursor on the context and changes the previous_requested_subset from a property to a function that takes the asset key.

We still want to keep the `_cursor` around as a reference to the root target's cursor as other properties/methods use this in ways that do actually want that cursor (e.g. getting the previous max storage id for an asset).

## How I Tested These Changes

Added test failed before, succeeds now.

## Changelog

Fixed issue with `AutomationCondition.newly_requested()` which could cause it to fail when nested within `AutomationCondition.any_deps_match()` or `AutomationCondition.all_deps_match()`.