[SelectionInput] Move `instance.setValue` into raf.  (#28560)
## Summary & Motivation
Theres a Heisenberg visual glitch if you try to set the value of the
codemirror instance in a layout effect. Since it's inconsistent I
suspect it's be a timing issue with their own internal layout effects so
moving the `setValue` call to a rAF which seems to consistently upgrade
the syntax visually.