Fix reconstruction metadata routing through DefinitionsLoadContext (#28416)
## Summary & Motivation

Fix a longstanding regression in reconstruction metadata routing that was introduced in #25436. Prior to that PR, reconstruction metadata during initial repository load was stored on `Definitions` objects in the `metadata` field. Integrations that used the reconstruction metadata API were expected to return a `Definitions` object that would be merged into the final `Definitions` object for the code location. This final `Definitions` object would then contained the merged reconstruction metadata for any integrations using the API:

```
my_integration_resource = MyIntegrationResource(...)
integration_defs = build_my_integration_defs(my_integration_resource)

# final defs object merges together metadata from all passed in defs objects
defs = Definitions.merge(
  integration_defs,
  # ... other definitions
)
```

#25436 aimed to eliminate the restriction that integrations using reconstruction metadata needed to return `Definitions` objects, so that they could return e.g. `AssetSpecs` without any wrapper. It did this by switching the channel by which reconstruction metadata was routed from `Definitions.metadata` to `DefinitionsLoadContext`. New methods `DefinitionsLoadContext.{add,get}_pending_reconstruction_metadata` were added. The idea was that integrations would stash their reconstruction metadata on a process-scoped `DefinitionsLoadContext`. Then the final bundle of reconstruction metadata would be pulled off the global `DefinitionsLoadContext` where previously it had been pulled off the final `Definitions` object:

```
my_integration_resource = MyIntegrationResource(...)
# build_my_integration_assets modified globally available `DefinitionsLoadContext`
integration_assets = build_my_integration_assets(my_integration_resource)

# final defs object merges together metadata from all passed in defs objects
defs = Definitions.merge(
  assets=my_integration_assets,
  # ... other definitions
)
```

There was a problem-- the `DefinitionsLoadContext` object was not actually process-scoped. `DefinitionsLoadContext.get()` would return a process-scoped singleton IF `DefinitionsLoadContext.set()` had been called first. But `DefinitionsLoadContext.set()` was NOT being called as part of code server startup, because previously it didn't need to be, since we weren't persisting data on the `DefinitionsLoadContext` during the initialization phase. Consequently, `add_pending_reconstruction_metadata` was being called on a throwaway instance of `DefinitionsLoadContext`-- this silently defeated the entire purpose of the system, which is to retrieve reconstruction metadata once at code server startup and then reuse that metadata in spawned run/step processes.

The reason this wasn't caught was inadequate testing (which to be clear was the situation prior to #25436). It's not that we had _no_ test coverage, but rather than our test environment was not quite the same as the real environment. Testing that reconstruction metadata for an external integration is appropriately cached cross-process is tricky-- we need to check that an external API is only hit once across multiple processes. The tests we had in place did not do this. Instead the external API was mocked and we conducted a simulated in-process initialization and reconstruction of a repo and execution (see `test_state_derived_defs` in `dagster-powerbi` for an example).

Routing of reconstruction metadata worked on this codepath but not the real-world code path of code server -> run/step processes. This codepath continued to execute without crashing, but cached reconstruction metadata was not being passed, instead being silently refetched in each process.

A final wrinkle is that another PR from a separate author/workstream that manipulated the reconstruction metadata system, #24632, landed one day after #25436, without being aware of the changes made to metadata routing in the latter PR. But it actually worked because the old metadata routing system wasn't totally removed in 25436-- it was partially removed, but then part of it was added back in yet another PR, #25483.

The morals of this long and complicated story are to (a) not cut corners on real-life cross-process testing; (b) when replacing an old internal system with a new one, totally remove the old system.

A worthy followup would make `DefinitionsLoadContext.get()` return a singleton unless that singleton is explicitly erased. A problem with this is that during code server startup we actually use `ReconstructableRepository` after initial load (I don't fully understand why yet), leading to the need to construct a different `DefinitionsLoadContext` in "reconstruction" mode after the initialization phase of the code server process.

## How I Tested These Changes

Added a new test (in `dagster-graphql`) that spins up an actual real-life code server and run /step processes and ensures that reconstruction metadata is appropriately cached.

Also manually tested that the API is hit only once for `dagster-fivetran`, using the repro in this thread.

I did not adjust all of our integrations using reconstruction metadata to use a real-life code server cross-process test, because this would've required modifying those integrations to perform some kind of logging that could be detected cross-process. I think the test I added (which ensures reconstruction metadata gets communicated cross-process for a generic integration) coupled with the existing "simulated cross-process execution" tests for the real integrations are sufficient.

## Changelog

Fixed a bug with several integrations that caused data fetched from external APIs not to be properly cached during code server initialization, leading to unnecessary API calls in run and step worker processes. This affected: `dagster-airbyte`, `dagster-dlift`, `dagster-dbt`, `dagster-fivetran`, `dagster-looker`, `dagster-powerbi`, `dagster-sigma`, and `dagster-tableau`.