Ensure that run monitoring covers runs in a NOT_STARTED case
Summary:
This is a somewhat unexpected case, but currently runs are in a NOT_STARTED state briefly after being created, before they transition to QUEUED (Ideally they would move right into QUEUED in a transaction, but that would be a larger change).

Ensure that in the unlikely event that a run gets stuck in NOT_STARTED state that it is eventually cleaned up by run monitoring.

New test case

Fixed an issue where run monitoring sometimes didn't fail runs that were stuck in a NOT_STARTED status instead of a STARTING status.

> Insert changelog entry or delete this section.