[dg][refactor] Rework ComponentDeclNode system (#28558)
## Summary & Motivation

NOTE: I don't think the names I've chosen here (`DefsModule` and `DefsModuleDecl`) are the best, definitely down to bikeshed that in later PRs, mostly want to get this refactor in to set the groundwork for better behavior in the future.

I couldn't figure out a way to do this more incrementally, so this is a bit of a large diff.

Previously, the way the system worked was to:

* parse the root `defs/` directory to create a tree structure of `ComponentDeclNodes`
* take that tree and call `load()` on the root node, which would return a flat list of `Component` objects
* call `build_defs()` on each of those `Component` objects, and `Definitions.merge()` them all together

There's a few issues with this setup.

First, this requires us to create a Component for any new loading behavior we want. This led to the pretty-odd ImplicitDefinitionsComponent setup where we would automatically generate a DefinitionsComponent on the user's behalf to enable our auto-loading behavior.

Next, by creating a flat list of `Component` objects, we lose the ability to inject additional information (such as additional properties) into the `build_defs()` logic at non-component levels of the hierarchy. i.e. this new refactor was designed to allow us to tack on additional files / info to DefsModuleDecls that could be passed through to the resulting DefsModules in a straightforward way.

Finally (and most personally-annoying), we have a whole nest of latent bugs owing to the fact that the ComponentLoadContext that we pass to Components during the `build_defs()` process actually has the wrong path attached to it in any case where you have a component nested in a subfolder, and there was no easy way of fixing this in the old regime as the Component lost access to its original ComponentLoadContext / path as soon as it was loaded. Now we keep a pointer to the load context for the relevant DefsModule types which fixes a whole range of issues I was running into when trying to iron out kinks in the implicit definitions loading code.

High level, we now:

* parse the root `defs/` directory to create a tree structure of `DefsModuleDecl`s
* take that tree and call `load()` on the root node, which returns a tree structure of `DefsModule`s
* call `build_defs()` on the root node, which returns a `Definitions` object

This hands over much more of the responsibility to the actual tree structure, which makes it possible to inject fun stuff (such as the aformentioned post processors).

It is at least somewhat arbitrary what amount of "work" goes into the "DefsModuleDecl" vs. the full "DefsModule", but my basic organization scheme is:

1. you should be able to instantiate all DefsModuleDecls without reading the contents of any files. You just analyze the available files in each directory and use that to construct your tree of declarations. This is why the logic relies on filenames (e.g. component.yaml, component.py, definitions.py)
2. you should be able to instantiate each DefsModule without explicitly instantiating any Definitions objects, as build_defs() handles that

it's a bit of a fuzzy line but we can push things around if/when we find more use for the `DefsModuleDecl` tree (right now we basically just instantly call load() on it, so the in-memory object isn't very useful outside of code-organization).


## How I Tested These Changes

Existing tests, also added some more implicit defs loading tests that failed in the previous regime.

## Changelog

NOCHANGELOG