[ui] Allow canceling job backfills after all runs are queued (UI only) (#28626)
## Summary & Motivation

This is discussed at length in
https://github.com/dagster-io/dagster/pull/28161 - this is just the
front-end portion of the work. Removing the graphQL resolver can come in
the next release after this has baked so that the UI and backend don't
need to ship at the same time.

## How I Tested These Changes

Canceled a backfill which had finished queueing runs.

## Changelog

[ui] Fixes a regression in the "Cancel Backfill" option for job
backfills which have finished queuing runs.

Co-authored-by: bengotow <bgotow@elementl.com>