AssetMaterializationFailure class (#28568)
## Summary & Motivation
Small refactor of the `AssetFailedToMaterializeData` class to accept an
`AssetMaterializationFailure` class. This class is like the
`AssetMaterialization` and `AssetObservation` classes. We'll need the
`AssetMaterializationFailure` class in order to report runless failed
materializations (ie for airlift) and it makes some of the GQL changes
in the stacked PR follow existing convention a bit better

internal pr https://github.com/dagster-io/internal/pull/14413

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.