[dagster-dbt] Improve default ad hoc job name for DbtCloudWorkspace (#28552)
## Summary & Motivation

Fixes AD-949.

This PR uses the project name and environment name, if available, to
create the ad hoc job. We fallback to project ID and environment ID if
the names are not returned in the API responses - the name field is not
required in both responses.

## How I Tested These Changes

Updated tests with BK. 

Tested locally with a live instance:
<img width="504" alt="Screenshot 2025-03-17 at 6 15 18 PM"
src="https://github.com/user-attachments/assets/d148f6cf-1a39-4d13-9ea7-727916130fb5"
/>