[ui] Allow toast actions to be links (#28638)
## Summary & Motivation

We received user feedback that with the new toasts, kicking off a materialization does not currently allow command-click to view the launched run in a new tab. This is a regression from previous toasts, which just rendered a link when an href was provided.

To resolve this, allow `showSharedToaster` to receive a `custom` action, which allows rendering a custom element in the toast. This lets callsites specify an `AnchorButton` to enable react-router-based navigation on click, or command-click to open the link in a new tab.

This required moving `GlobalStyleProvider` under the `BrowserRouter` provider in the `AppProvider` root, in order to enable react-router links to be rendered within toasts. This seems safe to me, since `GlobalStyleProvider` is mostly responsible for injecting global stylesheets anyway, so no providers higher up the tree should really be affected.

## How I Tested These Changes

Materialize an asset. Verify that I can click the resulting toast to view the run within the current tab, or command-click to open it in a new tab.

## Changelog

[ui] Allow using command-click to view a run from the toast message that appears when starting a materialization of an asset.