[ez] When merging external assets from multiple code locations, prefer non-stub asset specs (#28639)
## Summary

Right now, when building the global asset graph, we merge multiple
copies of the same asset into a canonical definition.

This generally prefers executable assets to specs, but provides no
disambiguation if the only representations for an asset are
non-executable specs.

This can lead to a poor UX if, for example, a user adds a
kind-and-description-augmented spec to one code location and in another
location has a spec autogenerated for an upstream dependency. It's very
possible for the bare, auto-generated spec to take precedence.

This PR de-prioritizes autogenerated specs in this disambiguation logic.

## Changelog

> When merging assets from multiple code locations, autogenerated specs
are now prioritized lower than customized external asset specs.