[time-window-partitions][fix] Fix calculation of num_partitions for subsets with negative ranges (#28637)
## Summary & Motivation

Still investigating exactly how this can happen -- it seems to be something where the `AllPartitionsSubset` behaves a bit strangely when handling certain intersection operations, but we're gonna need this code to handle objects that were serialized before the fix anyway.

## How I Tested These Changes

## Changelog

Fixed an issue that could cause automation condition evaluations to fail to render in the UI in rare cases.