[asset-backfill] Fix bug when target subset contains empty PartitionsSubset (#28608)
## Summary & Motivation

When you create an AssetBackfill from the point of failure, and the
original root asset had all of its partitions succeed, this creates a
target AssetGraphSubset that looks like:


```python
AssetGraphSubset(
    non_partitioned_keys=[...],
    partitions_subsets_by_key={
        AssetKey("root"): DefaultPartitionsSubset([]),
        ...
    }
)
```

That's totally fine and expected! And the asset backfill code mostly
graciously handles finding the new roots and all that. The issue is this
line:
https://sourcegraph.com/github.com/dagster-io/dagster/-/blob/python_modules/dagster/dagster/_core/execution/asset_backfill.py?L239

Where we check if we have successfully found all of the partitions in
the target set by mapping from the root partitions. The thing is... we
have! It's just that the target set has that `AssetKey("root"):
DefaultPartitionsSubset([])` in it, whereas the
root_and_downstream_partitions set does not even have a key for that
dictionary, so technically those dictionaries are different.

So the solution is to just filter out the empty subsets from the
dictionaries before comparing them. The foolproof thing would be to
calculate all the asset partition keys explicitly, but that's way too
expensive to do in practice, and this should cover the edge case well.

Note that you also have to override `__ne__` to handle the != operator,
that doesn't just default to "not __eq__"

## How I Tested These Changes

## Changelog

Fixed an issue that could cause an asset backfill created by
re-executing another backfill from the point of failure to error on the
first tick in rare cases.