[dg] DefsModuleComponent (#28660)
## Summary & Motivation

This is an alternative version of
https://github.com/dagster-io/dagster/pull/28624, in which we create a
wrapper component type instead of a new config file format.

The functionality is generally the same -- I copied over the exact test
from the other PR, and just updated the files:

Before:

```yaml
asset_post_processors:
  - target: "*"
    attributes:
      tags:
        foo: "bar"
```

After:

```yaml
type: dagster_components.dagster.DefsModuleComponent

attributes:
  asset_post_processors:
    - target: "*"
      attributes:
        tags:
	  foo: "bar"
```

(and renaming from defs.yaml to component.yaml)

The main differences that I think make this a bit less appealing
long-term:

1. It can't be used inside a single component's directory (because it
itself is a component)
2. Right now, a component conceptually "takes over" the loading of an
entire directory. This has implications for things like defs module
caching, and future worlds in which we want to render the tree structure
of the decl nodes without having to load all of the nodes completely. In
my code, I try to make the loading behavior as similar as possible to
what it would be without this component existing there (that's why I
hijack the `load()` method on the component class -- I want to make sure
that the load method for submodules gets called at the same time as it
would otherwise, as opposed to (e.g.) calling it inside the `build_defs`
function.

That being said, the new `defs.yaml` file is a bit more disruptive than
this in the short term, so we can go with this for now

## How I Tested These Changes

## Changelog

Added a new `dagster_components.dagster.DefsModuleComponent` that can be
used at any level of your `defs/` folder to apply asset attributes to
the definitions at or below that level.