[dg] Allow arbitrary objects to be represented as LibraryObjects (#28395)
## Summary & Motivation

This makes it possible for non-component RemoteLibraryObjects to exist in `dagster-dg`, and renames the dagster-components (internal) command from `list component-types` to `list library-objects`. This also changes how we detect "library objects". Previously, we just looked for Component subclasses, but now we scoop up anything with the `__dg_library_object__` attribute set, and represent that properly all the way through dagster-dg

This is a larger set of changes, mostly because we need to synchronize all the dagster-dg changes with corresponding dagster-components changes.

I've reworked the shape of the metadata that we serialize and send over the wire to be just:

```
{
  "object": {...},
  "scaffolder": {...},
}
```

, with the first field being required (representing metadata about the object), and the second being optional (not all objects can be scaffolded). Previously, we had a fairly-confusing layering of TypedDicts going on. I think this setup is easier to understand and more flexible for the future (e.g. if we want to add more types of objects than Component types), hence the more generic object_metadata / object_type setup.

Any object can have a scaffolder, so that remains a separate field.

NOTE: despite this change, there's currently no mechanism to set the magic attribute on anything other than component types. this will be done in the next PR, which will update the `with_scaffolder` method to set the proper attribute on its target.

## How I Tested These Changes

## Changelog

NOCHANGELOG