Update "Asset versioning and caching" guide (#28672)
## Summary & Motivation

Updates the "Asset versoining and caching guide" with new screenshots
and "Unsynced" terminology.

Note I was unable to update the last section in this PR due to an
apparent bug with observable source assets. After that is addressed I
will update the final section.

## How I Tested These Changes

Stepped through the guide.

## Changelog

[docs] The "Asset versioning and caching" guide has been updated to
reflect the current Dagster UI and "Unsynced" labels.