[docs] hard code names of GraphQL API Python client method (#28085)
## Summary & Motivation
Thee `PyObject` that links to [the API docs for the
GraphQL Python
Client](https://docs.dagster.io/api/python-api/libraries/dagster-graphql#python-client)
does not display the name of the client methods.

Changing `method` to `module` will show the name but the link will 404.

This PR hard codes in
the method names and links to their sections.

Before:
![Screenshot 2025-02-27 at 00 28
43](https://github.com/user-attachments/assets/86a11f91-8439-423f-9a84-a203f0241d84)

After:
![Screenshot 2025-02-27 at 00 29
01](https://github.com/user-attachments/assets/f01e02fa-931d-49b0-9f2e-6b4932b2d406)

## How I Tested These Changes
👀