[da] Update initial_evaluation condition to only detect partitions_def changes (#28764)
## Summary & Motivation

This is actually more aligned with the purpose of this condition, as you could imagine having an unpartitioned asset with `AutomationCondition.eager()` being turned into a static-partitioned asset with 100s of partitions and experiencing the bad behavior of all partitions updating.

It's also more accurate that this is the "initial_evaluation" of those partitions in this case.

## How I Tested These Changes

## Changelog

The `AutomationCondition.initial_evaluation` condition has been updated to become true for all partitions of an asset whenever the PartitionsDefinition of that asset changes, rather than whenever the structure of the condition changes.