[dagster-dbt] Add list_run_artifacts to DbtCloudWorkspaceClient (#28730)
## Summary & Motivation

This PR adds the list_run_artifacts method that will be needed in the
polling sensor. We need this because not all runs will have a
`run_results.json` artifact, so we'll want to request and verify that it
is in the list of artifacts for a given run before fetching it.

## How I Tested These Changes

New tests with BK