[ui] Restore __typename to tag filtering on automations (#28806)
## Summary & Motivation

In automation tag filtering, we're depending on the `__typename` key/value pair that I discarded in https://github.com/dagster-io/dagster/pull/28583. Tag filtering is currently broken because of its removal. Revert that change.

I think I missed this because the tag set that I was checking appeared to result in an expected empty set.

## How I Tested These Changes

View automations with a tag filter, verify that it works correctly. Add multiple tags to filter, verify that this continues to work correctly.

## Changelog

[ui] Fix tag filtering on automations list.