[ui] Modify arrayLimit on qs.parse for useQueryPersistedState (#28799)
## Summary & Motivation

Modify `qs.parse` usage in `useQueryPersistedState` to override the `arrayLength` default of 20 items. `qs.parse` will currently convert an array of 20+ items into an object, which breaks callsites that expect an encoded array.

The limit is raised to (a fairly arbitrary) 1000 items, though it is unlikely that we will try to encode so many items anywhere in our app, and the GET request querystring will be limited by length regardless. If the limit is breached, call `console.error` to track it internally for investigation, and allow the parse to proceed so that the user is not blocked.

## How I Tested These Changes

Added jest tests.

View global asset graph, command-click a whole bunch of assets in the sidebar. Verify that the URL updates correctly and is parsed properly.

## Changelog

[ui] Repair an issue with querystring parsing that can arise when selecting a large number of items the in selection syntax input.