update model fields access (#28808)
fix up some callsite where we are calling `model_fields` on the instance
instead of the class generating new warnings

```
pydantic.warnings.PydanticDeprecatedSince211: Accessing this attribute on the instance is deprecated, and will be removed in Pydantic V3. Instead, you should access this attribute from the model class. Deprecated in Pydantic V2.11 to be removed in V3.0.
```

## How I Tested These Changes

bk