[ui] Fix colorblind color mapping in Insights (#28812)
## Summary & Motivation

The mapping on dataviz `green` colors goes to a nonexistent color in colorblind themes, which leads to a JS error when we try to extract the color for Insights line charts. Fix this by mapping to a core cobalt color instead.

Also clean up a couple of unused color names.

## How I Tested These Changes

Switch to system colorblind theme, view Insights. Verify that the chart loads, with no JS errors.

## Changelog

[ui] Fix line charts for colorblind themes.