[ui] Add “Copy Tags” option to tags modals, unify copy behavior
related: https://linear.app/dagster-labs/issue/FE-637/add-button-to-copy-all-tags-in-tags-modal

This PR adds a “Copy Tags” button to the run tags modal and the run “view configuration” modal  (which also shows the tags).

In building this I realized we have a handful of copy buttons and their behaviors differ slightly, so I consolidated them down to two - an icon and a standard button. We could alternatively keep these different verisons and push more of the toaster behavior into the hook, but I think it’s nice to standardize on the icon appearance a bit too.