[ui] Add “Copy tags” option to tags modals, unify copy behavior (#28841)
related:
https://linear.app/dagster-labs/issue/FE-637/add-button-to-copy-all-tags-in-tags-modal

## Summary & Motivation

This PR adds a “Copy Tags” button to the run tags modal and the run
“view configuration” modal (which also shows the tags) which copies all
the tags as a yaml block.

In building this I realized we have a handful of copy buttons and their
behaviors (icons, toasts, etc) differ slightly, so I consolidated them
down to two - an icon and a standard button. We could alternatively keep
different buttons and push more toaster behavior into the hook, but I
think it’s nice to standardize on a single icon + behavior. The version
in the asset header is quite nice and I moved them all to that, but with
the larger click target.

There is one that I didn't touch, which is pulled in from dagster dg -
maybe we move this component to the new shared package and use it there
too 👀

## How I Tested These Changes


![image](https://github.com/user-attachments/assets/d68f9411-060b-4f47-8b00-7c5b729de749)

![image](https://github.com/user-attachments/assets/98171e96-5b06-4a3d-9696-2a0b8d55e7a0)

## Impacted


![image](https://github.com/user-attachments/assets/b8cbd4a7-648a-4c64-b73c-b9fced7e8925)

![image](https://github.com/user-attachments/assets/7538424d-11ac-413b-a565-4d4e843086ab)

![image](https://github.com/user-attachments/assets/25e2ba24-3da1-4f86-b5c1-b1632a2cc1cb)

## Changelog

[ui] Dagster's UI now allows you to copy run tags as a yaml block from
the Tags and Configuration modals

Co-authored-by: bengotow <bgotow@elementl.com>