[ui] Fix backfill toast URL (#28884)
## Summary & Motivation

Now that toasts are within the scope of the top-level `BrowserRouter`, they no longer need special link behavior, and `history.createHref` will in fact double the base path in the URL if used in conjunction with react-router links like `AnchorButton`.

This change fixes broken toast links when launching backfills.

## How I Tested These Changes

Launch a backfill in cloud, verify that the link in the resulting toast goes to the correct backfill page.

## Changelog

[ui] Fix link in toast messages that appear when launching backfills.