perf improvements when fetching partition ranges in get_2d_run_length_encoded_partitions (#28874)
## Summary & Motivation
The fact that each of these calls had a different current_timestamp and potentially invoke get_last_partition_window. which can be surprisingly expensive if the cronstring is unusual,was making this call unneccesarily slow for certain partitions definitions.

## How I Tested These Changes
run a slow dry run script of PartitionHealthQuery on a 2D multipartitions definition that was slow before, now is fast.

## Changelog
Performance improvements for loading the partitions page for multi-partitioned assets.