[dg] Remove DefsModuleDecl (#28779)
## Summary & Motivation

The DefsModuleDecl layer is not necessary, and this PR proves that by removing it. This also removes the vestigial DefinitionsModuleCache.

Instead, we modify the loading process for the `defs/` directory by invoking `DefsModuleComponent.from_context()` at the top level of the directory. This scans the directory, building up a tree of DefsModuleComponents, which can be one of:

`DefsFolderComponent`: This represents a folder in the hierarchy

`DagsterDefsModule`: A python file or directory that contains dagster definitions (assets, schedules, sensors, etc.) that will be automatically loaded

`WrappedDefsModule`: This represents an explicit component definition (pythonic or yaml-based).

This results in a pretty large simplification of the codebase. I also updated / simplified a bunch of tests that previously were manually constructing YamlComponentDecls.

## How I Tested These Changes

## Changelog

NOCHANGELOG