[dg] Remove DefinitionsComponent, DefsModuleComponent (#28811)
## Summary & Motivation

This deletes the `DefinitionsComponent` and repurposes the `FolderDefsModule` to serve the role of the `DefsFolderComponent` (i.e. unifies them both into a single class).

In essence, previously by default we would automatically generate a `DefsFolderComponent` at each layer of the hierarchy for you, but this component was not possible to configure. Now, I've added post processors to that config schema, and exported this from the dagster-components library.

We have to do some mildly funky stuff in the load path because this component actually is highly-dependent on the ComponentLoadContext in order to populate its resolved fields, but it's really not that bad (Resolved is a great framework!)

Similar treatment can be applied to the other core classes, and arguably I should do the same thing for the DagsterDefinitionsComponent instead of deleting it outright.

## How I Tested These Changes

## Changelog

TODO